+++ src/js/app.js
@@ -1,0 +1,7 @@
+new Vue({
+    el: "#app",
+    components:
+    {
+        "hs-goals": require('./goalsComponent/main.js'),
+    }
+});
