--- src/js/main.js
@@ -1,3 +1,0 @@
-new Vue({
-    el: "#app"
-});
